+++ src/vns/tb2dgvns.cpp
@@ -142,7 +142,7 @@ bool VNSSolver::solve()
                             k *= 2;
                             break;
                         case VNS_LUBY:
-                            k = ToulBar2::vnsKmin * luby(rank);
+                            k = ToulBar2::vnsKmin * (int)luby(rank);
                             break;
                         case VNS_ADD1JUMP:
                             if (ToulBar2::vnsNeighborVarHeur == RANDOMVAR || k < ((ClustersNeighborhoodStructure*)h)->getMaxClusterSize() + ((ClustersNeighborhoodStructure*)h)->getSize() - 1)
@@ -186,7 +186,7 @@ bool VNSSolver::solve()
                         lds *= 2;
                         break;
                     case VNS_LUBY:
-                        lds = ToulBar2::vnsLDSmin * luby(restart);
+                        lds = ToulBar2::vnsLDSmin * (int)luby(restart);
                         break;
                     default:
                         cerr << "Unknown LDS increment strategy inside VNS (see option -linc)!" << endl;
